requirements/gcp_requirements.txt
google-cloud-compute==1.10.0
google-cloud-bigquery==3.11.0
pyyaml==6.0